Cleanup loading of RequiredBlockEntrys
